% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{pdfetch}
\alias{pdfetch}
\alias{pdfetch-package}
\alias{pdfetch-package}
\title{A package for downloading economic and financial time series from public sources.}
\description{
A package for downloading economic and financial time series from public sources.
}
