% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfetch.R
\name{pdfetch_BUNDESBANK}
\alias{pdfetch_BUNDESBANK}
\title{Fetch data from the Deutsche Bundesbank}
\usage{
pdfetch_BUNDESBANK(identifiers)
}
\arguments{
\item{identifiers}{a vector of series codes}
}
\value{
a xts object
}
\description{
Fetch data from the Deutsche Bundesbank
}
\examples{
\dontrun{
pdfetch_BUNDESBANK(c("BBNZ1.Q.DE.Y.H.0000.A","BBK01.BJ9069"))
}
}
\seealso{
\url{https://www.bundesbank.de/Navigation/EN/Statistics/Time_series_databases/time_series_databases.html}
}
