% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfetch.R
\name{pdfetch_ECB}
\alias{pdfetch_ECB}
\title{Fetch data from European Central Bank's statistical data warehouse}
\usage{
pdfetch_ECB(identifiers)
}
\arguments{
\item{identifiers}{a vector of ECB series IDs}
}
\value{
a xts object
}
\description{
Fetch data from European Central Bank's statistical data warehouse
}
\examples{
\dontrun{
pdfetch_ECB("FM.B.U2.EUR.4F.KR.DFR.CHG")
}
}
\seealso{
\url{http://sdw.ecb.europa.eu/}
}

