% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfetch.R
\name{pdfetch_EUROSTAT_DSD}
\alias{pdfetch_EUROSTAT_DSD}
\title{Fetch description for a Eurostat dataset}
\usage{
pdfetch_EUROSTAT_DSD(flowRef)
}
\arguments{
\item{flowRef}{Eurostat dataset code}
}
\description{
Fetch description for a Eurostat dataset
}
\examples{
\dontrun{
pdfetch_EUROSTAT_DSD("namq_10_gdp")
}
}
