% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfetch.R
\name{pdfetch_WB}
\alias{pdfetch_WB}
\title{Fetch data from World Bank}
\usage{
pdfetch_WB(indicators, countries = "all")
}
\arguments{
\item{indicators}{a vector of World Bank indicators}

\item{countries}{a vector of country identifiers, which can be 2- or
3-character ISO codes. The special option "all" retrieves all countries.}
}
\value{
a xts object
}
\description{
Fetch data from World Bank
}
\examples{
\dontrun{
pdfetch_WB("NY.GDP.MKTP.CD", c("BR","MX"))
}
}
\seealso{
\url{https://data.worldbank.org/}
}
