\name{pdfetch_EUROSTAT_DSD}
\alias{pdfetch_EUROSTAT_DSD}
\title{Fetch description for a Eurostat dataset}
\usage{
pdfetch_EUROSTAT_DSD(flowRef)
}
\arguments{
  \item{flowRef}{Eurostat dataset code}
}
\description{
Fetch description for a Eurostat dataset
}
\examples{
pdfetch_EUROSTAT_DSD("namq_gdp_c")
}

