% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdfetch.R
\name{pdfetch_INSEE}
\alias{pdfetch_INSEE}
\title{Fetch data from the French National Institute of Statistics and Economic Studies (INSEE)}
\usage{
pdfetch_INSEE(identifiers)
}
\arguments{
\item{identifiers}{a vector of INSEE series codes}
}
\value{
a xts object
}
\description{
Fetch data from the French National Institute of Statistics and Economic Studies (INSEE)
}
\examples{
tryCatch(pdfetch_INSEE(c("000810635")),
   error = function(e) {},
   warning = function(w) {}
)
}

