% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdfetch.R
\name{pdfetch_ONS}
\alias{pdfetch_ONS}
\title{Fetch data from the UK Office of National Statistics}
\usage{
pdfetch_ONS(identifiers, dataset)
}
\arguments{
\item{identifiers}{a vector of ONS series codes}

\item{dataset}{ONS dataset name}
}
\value{
a xts object
}
\description{
Fetch data from the UK Office of National Statistics
}
\examples{
tryCatch(pdfetch_ONS(c("LF24","LF2G"), "lms"),
         error = function(e) {},
         warning = function(w) {}
)
}

