% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfetch.R
\name{pdfetch_ONS}
\alias{pdfetch_ONS}
\title{Fetch data from the UK Office of National Statistics}
\usage{
pdfetch_ONS(identifiers, dataset)
}
\arguments{
\item{identifiers}{a vector of ONS series codes}

\item{dataset}{optional ONS dataset name, only used if a time series is available in multiple datasets.}
}
\value{
a xts object
}
\description{
The ONS maintains multiple data products; this function can be used to
retrieve data from the Time Series Explorer, see \url{https://www.ons.gov.uk/timeseriestool}
}
\examples{
\dontrun{
pdfetch_ONS(c("LF24","LF2G"), "lms")
}
}

